Improve compatibility between boards with different processors

+++ EasyTransfer/examples/EasyTransfer_RX_Example/EasyTransfer_RX_Example.pde
@@ -6,8 +6,8 @@ EasyTransfer ET;
 struct RECEIVE_DATA_STRUCTURE{
   //put your variable definitions here for the data you want to receive
   //THIS MUST BE EXACTLY THE SAME ON THE OTHER ARDUINO
-  int blinks;
-  int pause;
+  int16_t blinks;
+  int16_t pause;
 };
 
 //give a name to the group of data
